+++ fpp_ptc/h_definition.f90
@@ -67,7 +67,7 @@ logical :: check_excessive_cutting =.true.
 logical :: switch_to_drift_kick =.true.
 integer :: m6lim =10
 integer :: faclim =1
-integer(2) :: old_integrator_init =1
+integer  :: old_integrator_init =1
 
 TYPE sub_taylor
      INTEGER j(lnv)
